Phase 2/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "human-agent-team-simulator-backend"
version = "0.1.0"
description = "Backend skeleton for a human-agent team simulator."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2,<2.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "PyYAML>=6.0.2,<7.0.0",
  "SQLAlchemy>=2.0.32,<3.0.0",
  "psycopg[binary]>=3.2.1,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
